For SDK 10, smaller clip area for headers
For #25

One largely noticeable issue with Gingerbread was that the clipRect for the header decorations was not clipping properly (not sure the exact reason), but by specifying the clip area to the actual area for the header when it is in view and clipping to the padding of the recycler view when not in view seems to have resolved this. 

Not entirely sure if this has other repercussions at this point and this is a somewhat quick and dirty solution, but I tested the sample app on 2.3.7 and 5.1.1 and did not see any issues. Still testing to see if there are any other issues with sdk 10.

Lemme know if you have any feedback or things you want me to test before merging!